Go server to listen to serial port
This configures and opens a serial port and logs incomming sensor
readings to the console. So far it's pretty boring. Main item of
interest here is I'm configuring the serial port device file using
exec.Command() to invoke the `stty` CLI tool. Then, I open the
serial port device with os.Open(). This is meant to work on macOS
and Linux without pulling in external dependencies beyond the Go
standard library.