implement basic hub functionality
This monitors a serial port, parses incoming sensor messages,
summarizes them with a 24-hour min/max temperature, and sends the
results to an IRC server. This is built to be used with an IRC
display bot (https://github.com/samblenny/irc-display-bot) and a
LoRa Greenhouse Monitor system with 2 sensors and a base station
(https://github.com/samblenny/lora-greenhouse-monitor).