improve ircbot reconnect logic
For some reason, systemd occasionally restarts my server with a log
message that indicates the server exited cleanly. That suggests
that one or more of my goroutines for managing serial/irc/web IO
are messing up their error handling and reconnect logic.

This change attempts to make the IRC scanner more robust with
better debug logging and a fix for a possible goroutine leak.