more context handling for clean shutdown
These changes should fix stuck goroutines and result in better
debug logging for what happens as the server is shutting down. If
I got this right, all the goroutines should be cancelable now, so
they will log a message and exit cleanly in the event of ctrl-c or
other similar causes.